baby-user relation

+++ app/Http/Controllers/BabyController.php
@@ -8,23 +8,30 @@ use Tymon\JWTAuth\Facades\JWTAuth;
 
 class BabyController extends Controller
 {
+    
     public function insert(Request $request)
     {
-        $baby = new Baby();
-        $baby-> name =      $request->name;
-        $baby-> birthdate = $request->birthdate;
-        $baby-> gender =    $request->gender;
+        $baby =      new Baby();
+        $baby->      name =      $request->name;
+        $baby->      birthdate = $request->birthdate;
+        $baby->      gender =    $request->gender;
+        $baby->      user_id =   $request->user_id;
         $baby->save();
         
 
-        return response()->json(["baby_data" => $baby]);
+        return response()->json(['success'=>'Baby added successfully',"baby_data" => $baby]);
     }
 
-    public function index(Request $request)
+
+    public function index($user_id)
     {
-        $babies = Baby::all();
-        return response()->json($babies);
+        if ($babies = Baby::where('user_id', '=',$user_id)->get()) {
+            return response()->json($babies);
+        } else {
+            return response()->json('no babies found');
+        }
     }
+
 
     public function show($id)
     {
@@ -32,22 +39,24 @@ class BabyController extends Controller
         return response()->json($baby);
     }
 
+
     public function delete($id)
     {
-        $baby = Baby::find($id);
-        $baby->delete();
-        return response()->json([$baby]);
+        if ($baby = Baby::find($id)) {
+            $baby->delete();
+            return response()->json(["Baby deleted successfully",'baby_data'=>$baby]);
+        } else {
+            return response()->json(['error'=>'Baby not found']);
+        }
     }
 
     public function update(Request $request, $id)
     {
-        $token = JWTAuth::getToken();
-        $user = JWTAuth::toUser($token);
-        $baby = Baby::find($id);
-        $baby-> name =      $request->name;
-        $baby-> birthdate = $request->birthdate;
-        $baby-> gender =    $request->gender;
-        $baby-> user_id =   $user->id;
+        $baby =      Baby::find($id);
+        $baby->      name =      $request->name;
+        $baby->      birthdate = $request->birthdate;
+        $baby->      gender =    $request->gender;
+        $baby->      user_id =   $request->user_id;
         $baby->save();
 
         return response()->json(["baby_data" => $baby]);
